+++ app/models/user.rb
@@ -14,6 +14,7 @@ class User < ActiveRecord::Base
   has_many :datasets, -> { where deleted: false }
   has_many :dataset_file_audits
   has_many :lists
+  has_many :tools
 
   # User Methods
 
@@ -23,6 +24,10 @@ class User < ActiveRecord::Base
 
   def all_viewable_datasets
     Dataset.current.with_viewer( self.id )
+  end
+
+  def all_tools
+    Tool.current.with_editor( self.id )
   end
 
   def avatar_url(size = 80, default = 'mm')