+++ db/migrate/20131211150818_create_tools.rb
@@ -1,0 +1,17 @@
+class CreateTools < ActiveRecord::Migration
+  def change
+    create_table :tools do |t|
+      t.string :name
+      t.text :description
+      t.string :slug
+      t.string :logo
+      t.integer :user_id
+      t.boolean :deleted, null: false, default: false
+
+      t.timestamps
+    end
+
+    add_index :tools, :slug, unique: true
+    add_index :tools, :user_id
+  end
+end